[refactor] useEffect를 상위 페이지 (home.tsx)으로 옮겨 전역으로 관리 및 하위 컴포넌트는 랜더링만 구현 (역할 분담)
