+++ Client/src/components/character/MainCard.tsx
@@ -1,19 +1,9 @@
-import { useEffect } from 'react'
-import useAccountStore from '../../stores/AccountStore'
 import { useCharacterStore } from '../../stores/CharacterStore'
 import Loading from '../ui/Loading'
 
 const MainCard = () => {
   const MainCharacter = useCharacterStore((state) => state.MainCharacter)
-  const loadProfileData = useCharacterStore((state) => state.loadProfileData)
-  const characterName = useAccountStore((state) => state.character)
   const profileLoading = useCharacterStore((state) => state.profileLoading)
-
-  useEffect(() => {
-    if (characterName) {
-      loadProfileData(characterName)
-    }
-  }, [characterName, loadProfileData])
 
   return (
     <div className='mt-[-20px] flex h-40 items-center justify-center'>
